+++ src/app/shared/awards-banner/awards-banner.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core'
+
+@Component({
+  selector: 'app-awards-banner',
+  templateUrl: './awards-banner.component.html',
+  styleUrls: ['./awards-banner.component.scss']
+})
+export class AwardsBannerComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit() {}
+}